+++ .github/scripts/dev-requirements.txt
@@ -1,5 +1,4 @@
 singularity-hpc
-pipelib
 requests
 beautifulsoup4
 packaging
