+++ .github/workflows/update-cache.yaml
@@ -1,6 +1,7 @@
 name: Update Cache
 
 on:
+  pull_request: []
   workflow_dispatch:
   schedule:
   - cron: 0 0 * * 3
@@ -15,56 +16,20 @@ jobs:
     - name: Create conda environment
       run: conda create --quiet -c conda-forge --name shpc spython
 
-    - name: Install shpc and Guts
+    - name: Derive BioContainers List
       run: |
         export PATH="/usr/share/miniconda/bin:$PATH"
         source activate shpc
         pip install git+https://github.com/singularityhub/singularity-hpc@main
-        pip install git+https://github.com/singularityhub/guts@main
-        pip install -r .github/scripts/dev-requirements.txt
+        python .github/scripts/get_biocontainers.py /tmp/biocontainers.txt
+        head /tmp/biocontainers.txt
 
-    - name: Update Cache
-      run: |
-        export PATH="/usr/share/miniconda/bin:$PATH"
-        source activate shpc
-        printf "python .github/scripts/update_biocontainers.py\n"
-        python .github/scripts/update_biocontainers.py
-        printf "python .github/scripts/calculate_frequency.py\n"
-        python .github/scripts/calculate_frequency.py
+    - name: Update Cache Action
+      uses: singularityhub/container-executable-discovery@main
+      with:
+        token: ${{ secrets.GITHUB_TOKEN }}
+        repo-letter-prefix: true
+        listing: /tmp/biocontainers/containers.txt
+        dry_run: ${{ github.event_name == 'pull_request' }}
 
-    - name: Checkout Update branch
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        BRANCH_AGAINST: main
-      run: |
-        printf "GitHub Actor: ${GITHUB_ACTOR}\n"
-        export BRANCH_FROM="update/cache-$(date '+%Y-%m-%d')"
-        git remote set-url origin "https://x-access-token:${GITHUB_TOKEN}@github.com/${GITHUB_REPOSITORY}.git"
-        git branch
-        git checkout -b "${BRANCH_FROM}" || git checkout "${BRANCH_FROM}"
-        git branch
-        git config --global user.name "github-actions"
-        git config --global user.email "github-actions@users.noreply.github.com"
-        git add counts.json
-        git add skips.json
-        git add quay.io/biocontainers/*/*
-        git status
-        if git diff-index --quiet HEAD --; then
-           export OPEN_PULL_REQUEST=0
-           printf "No changes\n"
-        else
-           export OPEN_PULL_REQUEST=1
-           printf "Changes\n"
-           git commit -a -m "Automated deployment with new container $(date '+%Y-%m-%d')"
-           git push origin "${BRANCH_FROM}"
-        fi
-        echo "OPEN_PULL_REQUEST=${OPEN_PULL_REQUEST}" >> $GITHUB_ENV
-        echo "PULL_REQUEST_FROM_BRANCH=${BRANCH_FROM}" >> $GITHUB_ENV
-        echo "PULL_REQUEST_TITLE=[bot] ${BRANCH_FROM}" >> $GITHUB_ENV
-        echo "PULL_REQUEST_BODY=Add new container pull request." >> $GITHUB_ENV
-    - name: Open Pull Request
-      uses: vsoch/pull-request-action@50f22f6d146226ee6b73b7a001f26a3d4579f360 # 1.0.22
-      if: ${{ env.OPEN_PULL_REQUEST == '1' }}
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        PULL_REQUEST_BRANCH: main
+
